services for user moved to app.services, database and session ready
