+++ app/api/v1/users.py
@@ -1,10 +1,10 @@
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
 
-from app.api.v1.service import create, get_by_email
 from app.core.security import create_access_token
 from app.db.session import get_db
 from app.schemas.user import UserCreate, UserLogin, UserLoginResponse, UserOut
+from app.services.user import create, get_by_email
 
 user_router = APIRouter(prefix="/api/users", tags=["Users"])
 
@@ -33,4 +33,3 @@ def login_user(user_in: UserLogin, db: Session = Depends(get_db)):
 
     token = create_access_token(data={"sub": user.email})
     return {"token": token, "email": user.email}
-
