+++ app/core/config.py
@@ -5,3 +5,5 @@ config = Config(".env")
 JWT_SECRET = config("JWT_SECRET", default=None)
 JWT_ALG = config("JWT_ALG", default="HS256")
 JWT_EXP = config("JWT_EXP", cast=int, default=86400)
+
+DATABASE_URL = config("DATABASE_URL")
