--- app/db/base.py
@@ -1,3 +1,0 @@
-from sqlalchemy.ext.declarative import declarative_base
-
-Base = declarative_base()
