+++ app/db/database.py
@@ -1,0 +1,13 @@
+from sqlalchemy.ext.asyncio import (AsyncSession, async_sessionmaker,
+                                    create_async_engine)
+from sqlalchemy.orm import declarative_base
+
+from app.core.config import DATABASE_URL
+
+engine = create_async_engine(DATABASE_URL, echo=True, future=True)
+
+AsyncSessionLocal = async_sessionmaker(
+    bind=engine, class_=AsyncSession, expire_on_commit=False
+)
+
+Base = declarative_base()
