+++ app/db/session.py
@@ -1,0 +1,9 @@
+from contextlib import asynccontextmanager
+
+from app.db.database import AsyncSessionLocal
+
+
+@asynccontextmanager
+async def get_db():
+    async with AsyncSessionLocal() as session:
+        yield session
